Handle alpha value when creating diff graylevel mask
Pixmap one and two are now important no longer commutative. When pixmap
one is transparent, it should not be considered in the mask (=> black)
as we do not want to fill this space. If pixmap two is transparent, we
want it to be of high importance (=> white), as we want to fill this
space with high priority. So pixmap one should contain the target, while
pixmap two should contain the approximation.