Fix sample_random_xy conversion
The basis of the sampling comes from the data stored in tiny_skia
Pixmap, which returns a 1 dimensional array. However, the layout is such
that the width is the stride. Meaning, given a 100x200 image, the
indices 0-99 correspond to coordinates (0, 0..99).

Because of this the division has to use height, not width.